fix: gitlab plugins build
