+++ modulesInfo.go
@@ -1,5 +1,14 @@
 package velox
 
+import (
+	"fmt"
+	"regexp"
+	"strconv"
+	"time"
+
+	m "golang.org/x/mod/module"
+)
+
 // ModulesInfo represents single go module
 type ModulesInfo struct {
 	// Version - commit sha or tag
@@ -10,4 +19,27 @@ type ModulesInfo struct {
 	ModuleName string
 	// Replace (for the local dev)
 	Replace string
+}
+
+var vr = regexp.MustCompile("/v(\\d+)$") //nolint:gosimple
+
+// ParseModuleInfo here we accept a module name and return the version
+// e.g.: github.com/roadrunner-server/logger/v2 => v2
+func ParseModuleInfo(module string, t time.Time, rev string) string {
+	match := vr.FindStringSubmatch(module)
+	var version string
+	if len(match) > 1 {
+		if !IsDigit(match[1]) {
+			return m.PseudoVersion("", "", t, rev)
+		}
+
+		version = fmt.Sprintf("v%s", match[1])
+	}
+
+	return m.PseudoVersion(version, "", t, rev)
+}
+
+func IsDigit(num string) bool {
+	_, err := strconv.ParseInt(num, 10, 64)
+	return err == nil
 }
