+++ velox_rr_v2023.toml
@@ -20,7 +20,7 @@ ref = "v2023.3.0"
     lock = { ref = "v4.5.2", owner = "roadrunner-server", repository = "lock" }
 
     # CENTRIFUGE BROADCASTING PLATFORM
-    centrifuge = { ref = "v4.4.2", owner = "roadrunner-server", repository = "centrifuge" }
+    centrifuge = { ref = "v4.5.0", owner = "roadrunner-server", repository = "centrifuge" }
 
     # WORKFLOWS ENGINE
     temporal = { ref = "v4.5.0", owner = "temporalio", repository = "roadrunner-temporal" }
@@ -72,19 +72,19 @@ ref = "v2023.3.0"
     # TCP for the RAW TCP PAYLOADS
     tcp = { ref = "v4.3.2", owner = "roadrunner-server", repository = "tcp" }
 
-#[gitlab]
-#    [gitlab.token]
-#    # api, read-api, read-repo
-#    token = "${GL_TOKEN}"
-#
-#    [gitlab.endpoint]
-#    endpoint = "https://gitlab.com"
-#
-#    [gitlab.plugins]
-#    # ref -> master, commit or tag
-#    test_plugin_1 = { ref = "main", owner = "rustatian", repository = "36405203" }
-#    test_plugin_2 = { ref = "main", owner = "rustatian", repository = "36405235" }
-#
+[gitlab]
+    [gitlab.token]
+    # api, read-api, read-repo
+    token = "${GL_TOKEN}"
+
+    [gitlab.endpoint]
+    endpoint = "https://gitlab.com"
+
+    [gitlab.plugins]
+    # ref -> master, commit or tag
+    test_plugin_1 = { ref = "main", owner = "rustatian", repository = "36405203" }
+    test_plugin_2 = { ref = "main", owner = "rustatian", repository = "36405235" }
+
 [log]
 level = "debug"
 mode = "development"
